Projet2/requirement.txt
torch>=2.1.0
diffusers>=0.23.1
transformers>=4.33.0
gradio>=3.41.0
Pillow>=10.0.0
accelerate>=0.22.0
safetensors
